Initial commit - Streamlit SQL LLM chatbot

+++ streamlit_app.py
@@ -10,7 +10,7 @@ def load_model():
     model = AutoModelForCausalLM.from_pretrained(
         MODEL_NAME,
         device_map="auto",
-        torch_dtype=torch.float16,
+        dtype=torch.float32,
     )
     return tokenizer, model
